library/pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = ""
authors = [
    {name = "sachink108",email = "sachink108@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.102.0,<2.0.0)",
    "streamlit (>=1.49.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
